Move withoutFewShotExamples into shared catalog utilities
