+++ desktopApp/src/main/kotlin/ru/souz/tool/ToolsFactory.kt
@@ -2,10 +2,7 @@ package ru.souz.tool
 
 import ru.souz.agent.spi.AgentToolCatalog
 import ru.souz.db.SettingsProvider
-import ru.souz.llms.LLMRequest
-import ru.souz.llms.LLMResponse
 import ru.souz.llms.LLMToolSetup
-import ru.souz.llms.ToolInvocationMeta
 
 class ToolsFactory(
     runtimeToolCatalog: AgentToolCatalog,
@@ -26,20 +23,5 @@ class ToolsFactory(
                 }
             }
         ).toolsByCategory
-    }
-}
-
-private fun LLMToolSetup.withoutFewShotExamples(): LLMToolSetup {
-    val delegate = this
-    return object : LLMToolSetup {
-        override val fn: LLMRequest.Function = delegate.fn.copy(fewShotExamples = emptyList())
-
-        override suspend fun invoke(functionCall: LLMResponse.FunctionCall): LLMRequest.Message =
-            delegate.invoke(functionCall)
-
-        override suspend fun invoke(
-            functionCall: LLMResponse.FunctionCall,
-            meta: ToolInvocationMeta,
-        ): LLMRequest.Message = delegate.invoke(functionCall, meta)
     }
 }
