+++ sharedLogic/src/commonJvmMain/kotlin/ru/souz/tool/ToolCatalogs.kt
@@ -4,7 +4,10 @@ import java.util.Collections
 import ru.souz.agent.spi.AgentToolCatalog
 import ru.souz.db.SettingsProvider
 import ru.souz.llms.LLMChatAPI
+import ru.souz.llms.LLMRequest
+import ru.souz.llms.LLMResponse
 import ru.souz.llms.LLMToolSetup
+import ru.souz.llms.ToolInvocationMeta
 import ru.souz.llms.giga.toGiga
 import ru.souz.llms.runtime.ImageGenerationGateway
 import ru.souz.llms.runtime.VisionGateway
@@ -126,5 +129,20 @@ fun immutableToolCatalogSnapshot(
     val immutableSnapshot = Collections.unmodifiableMap(immutableCategories)
     return object : AgentToolCatalog {
         override val toolsByCategory: Map<ToolCategory, Map<String, LLMToolSetup>> = immutableSnapshot
+    }
+}
+
+fun LLMToolSetup.withoutFewShotExamples(): LLMToolSetup {
+    val delegate = this
+    return object : LLMToolSetup {
+        override val fn: LLMRequest.Function = delegate.fn.copy(fewShotExamples = emptyList())
+
+        override suspend fun invoke(functionCall: LLMResponse.FunctionCall): LLMRequest.Message =
+            delegate.invoke(functionCall)
+
+        override suspend fun invoke(
+            functionCall: LLMResponse.FunctionCall,
+            meta: ToolInvocationMeta,
+        ): LLMRequest.Message = delegate.invoke(functionCall, meta)
     }
 }
